+++ halotools/mock_observables/__init__.py
@@ -9,8 +9,6 @@ from .tpcf import *
 from .tpcf_jackknife import *
 from .tpcf_one_two_halo_decomp import *
 from .marked_tpcf import *
-from .rp_pi_tpcf import *
-from .wp import *
 from .group_identification import *
 from .mock_survey import *
 from .error_estimation_tools import jackknife_covariance_matrix
