+++ halotools/mock_observables/two_point_clustering/__init__.py
@@ -6,9 +6,12 @@ galaxy-galaxy lensing `~halotools.mock_observables.delta_sigma`, and more.
 """
 from __future__ import absolute_import 
 
-__all__ = ('angular_tpcf', 'delta_sigma', 's_mu_tpcf', 'tpcf_multipole')
+__all__ = ('angular_tpcf', 'delta_sigma', 's_mu_tpcf', 'tpcf_multipole', 'wp', 
+    'rp_pi_tpcf')
 
 from .angular_tpcf import angular_tpcf 
 from .delta_sigma import delta_sigma
 from .s_mu_tpcf import s_mu_tpcf
 from .tpcf_multipole import tpcf_multipole
+from .wp import wp
+from .rp_pi_tpcf import rp_pi_tpcf
