+++ halotools/mock_observables/two_point_clustering/tests/test_rp_pi_tpcf.py
@@ -1,44 +1,48 @@
-#!/usr/bin/env python
+""" Module provides unit-testing for the `~halotools.mock_observables.rp_pi_tpcf` function. 
+"""
+from __future__ import absolute_import, division, print_function, unicode_literals
 
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
 import numpy as np
-import sys
+from astropy.utils.misc import NumpyRNGContext
 
 from ..rp_pi_tpcf import rp_pi_tpcf
 
-__all__=['test_rp_pi_tpcf_auto_nonperiodic','test_rp_pi_tpcf_auto_periodic',\
-         'test_rp_pi_tpcf_cross_periodic','test_rp_pi_tpcf_cross_nonperiodic']
+__all__ = ('test_rp_pi_tpcf_auto_nonperiodic','test_rp_pi_tpcf_auto_periodic',
+    'test_rp_pi_tpcf_cross_periodic','test_rp_pi_tpcf_cross_nonperiodic')
 
 #create toy data to test functions
-Npts=100
-sample1 = np.random.random((Npts,3))
-sample2 = np.random.random((Npts,3))
-randoms = np.random.random((Npts,3))
 period = np.array([1.0,1.0,1.0])
 rp_bins = np.linspace(0,0.3,5)
 pi_bins = np.linspace(0,0.3,5)
+
+fixed_seed = 43
 
 def test_rp_pi_tpcf_auto_nonperiodic():
     """
     test rp_pi_tpcf autocorrelation without periodic boundary conditions
     """
+    Npts=100
+    with NumpyRNGContext(fixed_seed):
+        sample1 = np.random.random((Npts,3))
+        randoms = np.random.random((Npts,3))
+
     result = rp_pi_tpcf(sample1, rp_bins, pi_bins, sample2 = None, 
-                        randoms=randoms, period = None, 
-                        max_sample_size=int(1e4), estimator='Natural')
-    
-    print(result)
-    
+        randoms=randoms, period = None, 
+        max_sample_size=int(1e4), estimator='Natural')
+        
     assert result.ndim == 2, "More than one correlation function returned erroneously."
 
 def test_rp_pi_tpcf_auto_periodic():
     """
     test rp_pi_tpcf autocorrelation with periodic boundary conditions
     """
+    Npts=100
+    with NumpyRNGContext(fixed_seed):
+        sample1 = np.random.random((Npts,3))
     
     result = rp_pi_tpcf(sample1, rp_bins, pi_bins, sample2 = None, 
-                        randoms=None, period = period, 
-                        max_sample_size=int(1e4), estimator='Natural')
+        randoms=None, period = period, 
+        max_sample_size=int(1e4), estimator='Natural')
     
     assert result.ndim == 2, "More than one correlation function returned erroneously."
 
@@ -47,9 +51,14 @@ def test_rp_pi_tpcf_cross_periodic():
     """
     test rp_pi_tpcf cross-correlation without periodic boundary conditions
     """
+    Npts=100
+    with NumpyRNGContext(fixed_seed):
+        sample1 = np.random.random((Npts,3))
+        sample2 = np.random.random((Npts,3))
+
     result = rp_pi_tpcf(sample1, rp_bins, pi_bins, sample2 = sample2, 
-                        randoms=None, period = period, 
-                        max_sample_size=int(1e4), estimator='Natural')
+        randoms=None, period = period, 
+        max_sample_size=int(1e4), estimator='Natural')
     
     assert len(result)==3, "wrong number of correlations returned"
     assert result[0].ndim == 2, "dimension of auto incorrect"
@@ -61,9 +70,15 @@ def test_rp_pi_tpcf_cross_nonperiodic():
     """
     test rp_pi_tpcf cross-correlation without periodic boundary conditions
     """
+    Npts=100
+    with NumpyRNGContext(fixed_seed):
+        sample1 = np.random.random((Npts,3))
+        sample2 = np.random.random((Npts,3))
+        randoms = np.random.random((Npts,3))
+
     result = rp_pi_tpcf(sample1, rp_bins, pi_bins, sample2 = sample2, 
-                        randoms=randoms, period = None, 
-                        max_sample_size=int(1e4), estimator='Natural')
+        randoms=randoms, period = None, 
+        max_sample_size=int(1e4), estimator='Natural')
     
     assert len(result)==3, "wrong number of correlations returned"
     assert result[0].ndim == 2, "dimension of auto incorrect"
