+++ halotools/mock_observables/two_point_clustering/tests/test_wp.py
@@ -1,26 +1,30 @@
-#!/usr/bin/env python
+""" Module providing unit-testing of the `~halotools.mock_observables.wp` function.
+"""
+from __future__ import absolute_import, division, print_function, unicode_literals
 
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
 import numpy as np
-import sys
+from astropy.utils.misc import NumpyRNGContext
+
 from ..wp import wp
 
-__all__=['test_wp_auto_nonperiodic','test_wp_auto_periodic','test_wp_cross_periodic',\
-         'test_wp_cross_nonperiodic']
+__all__ = ('test_wp_auto_nonperiodic','test_wp_auto_periodic','test_wp_cross_periodic',
+    'test_wp_cross_nonperiodic')
 
-Npts = 100
-sample1 = np.random.random((Npts,3))
-sample2 = np.random.random((Npts,3))
-randoms = np.random.random((Npts,3))
 period = np.array([1.0,1.0,1.0])
 rp_bins = np.linspace(0,0.3,5)
 pi_max = 0.3
+
+fixed_seed = 43
 
 def test_wp_auto_nonperiodic():
     """
     test wp autocorrelation without periodic boundary conditions
     """
+    Npts = 100
+    with NumpyRNGContext(fixed_seed):
+        sample1 = np.random.random((Npts,3))
+        randoms = np.random.random((Npts,3))
+
     result = wp(sample1, rp_bins, pi_max, sample2 = None, 
                   randoms=randoms, period = None, 
                   max_sample_size=int(1e4), estimator='Natural')
@@ -33,6 +37,10 @@ def test_wp_auto_periodic():
     """
     test wp autocorrelation with periodic boundary conditions
     """
+    Npts = 100
+    with NumpyRNGContext(fixed_seed):
+        sample1 = np.random.random((Npts,3))
+
     result = wp(sample1, rp_bins, pi_max, sample2 = None, 
                 randoms=None, period = period, 
                 max_sample_size=int(1e4), estimator='Natural')
@@ -45,6 +53,11 @@ def test_wp_cross_periodic():
     """
     test wp cross-correlation with periodic boundary conditions
     """
+    Npts = 100
+    with NumpyRNGContext(fixed_seed):
+        sample1 = np.random.random((Npts,3))
+        sample2 = np.random.random((Npts,3))
+
     result = wp(sample1, rp_bins, pi_max, sample2 = sample2, 
                 randoms=None, period = period, 
                 max_sample_size=int(1e4), estimator='Natural')
@@ -58,6 +71,12 @@ def test_wp_cross_nonperiodic():
     """
     test wp cross-correlation with periodic boundary conditions
     """
+    Npts = 100
+    with NumpyRNGContext(fixed_seed):
+        sample1 = np.random.random((Npts,3))
+        sample2 = np.random.random((Npts,3))
+        randoms = np.random.random((Npts,3))
+
     result = wp(sample1, rp_bins, pi_max, sample2 = sample2, 
                 randoms=randoms, period = None, 
                 max_sample_size=int(1e4), estimator='Natural')
